requirements.txt
flask
pandas
numpy
tensorflow-cpu
scikit-learn
openpyxl